Move out common functionality to header file test_rocrand_prng
